Move everything from containers/ and pages/ into views/ and components/
