+++ imports/startup/client/routes.js
@@ -2,18 +2,18 @@ import React from 'react'
 import { IndexRoute, Router, Route, browserHistory } from 'react-router'
 
 // route components
-import AboutPage from '/imports/ui/pages/AboutPage'
-import AppContainer from '/imports/ui/containers/AppContainer'
-import HomePage from '/imports/ui/pages/HomePage'
-import LoginPage from '/imports/ui/pages/LoginPage'
-import NotFoundPage from '/imports/ui/pages/NotFoundPage'
-import OrganizationPage from '/imports/ui/pages/organizations/OrganizationPage'
-import OrganizationsPage from '/imports/ui/pages/organizations/OrganizationsPage'
-import PositionPage from '/imports/ui/pages/organizations/positions/PositionPage'
-import SearchPage from '/imports/ui/pages/SearchPage'
-import SettingsPage from '/imports/ui/pages/SettingsPage'
-import UserPage from '/imports/ui/pages/users/UserPage'
-import UsersPage from '/imports/ui/pages/users/UsersPage'
+import AboutPage from '/imports/ui/views/about/AboutPage'
+import AppContainer from '/imports/ui/components/AppContainer'
+import HomePage from '/imports/ui/views/home/HomePage'
+import LoginPage from '/imports/ui/views/login/LoginPage'
+import NotFoundPage from '/imports/ui/views/about/NotFoundPage'
+import OrganizationPage from '/imports/ui/views/organizations/OrganizationPage'
+import OrganizationsPage from '/imports/ui/views/organizations/OrganizationsPage'
+import PositionPage from '/imports/ui/views/organizations/positions/PositionPage'
+import SearchPage from '/imports/ui/views/search/SearchPage'
+import SettingsPage from '/imports/ui/views/settings/SettingsPage'
+import UserPage from '/imports/ui/views/users/UserPage'
+import UsersPage from '/imports/ui/views/users/UsersPage'
 
 export const Routes = () => (
   <Router history={browserHistory}>
